Refactor: Lotto클래스와 DrawnLotto 클래스에서 검증 부분 Validator클래스로 분리
